LayoutTests/imported/w3c/web-platform-tests/trusted-types/trusted-types-eval-reporting-expected.txt

CONSOLE MESSAGE: Refused to load because it does not appear in the object-src directive of the Content Security Policy.


FAIL Trusted Type violation report: evaluating a string. assert_throws_js: function "_ => eval('beacon="should not run"')" did not throw
PASS Trusted Type violation report: evaluating a Trusted Script.
FAIL Trusted Type violation report: default policy transforms the script before CSP checks runs. assert_equals: expected "default policy" but got "payload"
